fix(mobile): 노선 배지 접힘 개수 타입 고정

+++ apps/mobile/lib/station_search.dart
@@ -3458,7 +3458,7 @@ class StationLineBadges extends StatelessWidget {
     final maxCount = maxBadgeCount;
     final shouldCollapse = maxCount != null && lines.length > maxCount;
     final visibleLineCount = shouldCollapse
-        ? (maxCount - 1).clamp(1, lines.length)
+        ? (maxCount - 1).clamp(1, lines.length).toInt()
         : lines.length;
     final hiddenLineCount = lines.length - visibleLineCount;
